+++ R/project-template.R
@@ -1,0 +1,86 @@
+#' Create new SpaDES project
+#'
+#' Initialize a project with subdirectories \file{cache/}, \file{modules/},
+#' \file{inputs/}, \file{outputs/}, and \code{setPaths} accordingly.
+#'
+#' @param name project name (name of project directory)
+#' @param path path to directory in which to create the project directory
+#' @param ...  additional arguments (not used)
+#'
+#' @export
+#' @rdname newProject
+#'
+#' @examples
+#' myProjDir <- newProject("myProject", tempdir())
+#'
+#' dir.exists(file.path(myProjDir, "cache"))
+#' dir.exists(file.path(myProjDir, "inputs"))
+#' dir.exists(file.path(myProjDir, "modules"))
+#' dir.exists(file.path(myProjDir, "outputs"))
+#' unlink(myProjDir, recursive = TRUE) ## cleanup
+setGeneric("newProject", function(name, path, ...) {
+  standardGeneric("newProject")
+})
+
+#' @export
+#' @rdname newProject
+#' @importFrom reproducible checkPath
+setMethod(
+  "newProject",
+  signature = c(name = "character", path = "character"),
+  definition = function(name, path, ...) {
+    args <- list(...)
+
+    checkPath(path, create = TRUE)
+    projDir <- checkPath(file.path(path, name), create = TRUE)
+
+    setPaths(
+      cachePath = checkPath(file.path(path, name, "cache"), create = TRUE),
+      inputPath = checkPath(file.path(path, name, "inputs"), create = TRUE),
+      modulePath = checkPath(file.path(path, name, "modules"), create = TRUE),
+      outputPath = checkPath(file.path(path, name, "outputs"), create = TRUE)
+    )
+
+    ## TODO: check for Rstudio project?
+
+    newProjectCode(name, path)
+
+    return(projDir)
+})
+
+#' Create new module code file
+#'
+#' @param name  Character string specifying the name of the new module.
+#'
+#' @param path  Character string. Subdirectory in which to place the new module code file.
+#'              The default is the current working directory.
+#'
+#' @param open  Logical. Should the new project file be opened after creation?
+#'              Default \code{TRUE} in an interactive session.
+#'
+#' @author Alex Chubaty
+#' @export
+#' @rdname newProjectCode
+#'
+setGeneric("newProjectCode", function(name, path, open) {
+  standardGeneric("newProjectCode")
+})
+
+#' @export
+#' @importFrom reproducible checkPath
+#' @importFrom whisker whisker.render
+#' @rdname newModuleCode
+setMethod(
+  "newModuleCode",
+  signature = c(name = "character", path = "character", open = "logical"),
+  definition = function(name, path, open = interactive()) {
+    nestedPath <- checkPath(file.path(path, name), create = TRUE)
+    filenameR <- file.path(nestedPath, paste0(name, "-project.R"))
+
+    projectData <- list()
+    projectTemplate <- readLines(file.path(.pkgEnv[["templatePath"]],
+                                                     "project.R.template"))
+    writeLines(whisker.render(projectTemplate, projectData), filenameR)
+
+    if (open) .fileEdit(filenameR)
+})